Have persistent time format

+++ AlarmClock/AlarmClock/FirstViewController.swift
@@ -43,13 +43,17 @@ class FirstViewController: UIViewController {
         // get current time
         var date = NSDate();
         var format = NSDateFormatter();
-        format.dateFormat = "hh:mm:ssa";
+        var formatString: String;
+        
+        // set time format
+        if(self.militaryTime){
+            formatString = "HH:MM:ss";
+        }else{
+            formatString = "h:mm:ssa";
+        }
+        format.dateFormat = formatString;
+        
         var time = format.stringFromDate(date)
-        /**var components = calendar.components(.CalendarUnitHour | .CalendarUnitMinute | .CalendarUnitSecond, fromDate: date)
-        var hour = components.hour
-        var minutes = components.minute
-        var seconds = components.second
-        CurrentTime.text = "\(hour):\(minutes):\(seconds)"**/
         CurrentTime.text = time;
         
     }
@@ -58,16 +62,15 @@ class FirstViewController: UIViewController {
      * Loads saved alarm settings if they exist
      **/
     private func loadAlarm(){
+        // get entity model
         let entityDescription =
         NSEntityDescription.entityForName("Entity",
             inManagedObjectContext: managedObjectContext!)
         
+        // get results from coredata
         let request = NSFetchRequest()
         request.entity = entityDescription
-        
         var error: NSError?
-        
-        // get results from coredata
         var objects = managedObjectContext?.executeFetchRequest(request,
             error: &error)
         
@@ -77,27 +80,30 @@ class FirstViewController: UIViewController {
             var match:NSManagedObject;
             if results.count > 0 {
                 match = results[results.count - 1] as NSManagedObject
-                // prep date in specified format
-                var format = NSDateFormatter();
-                format.dateFormat = "hh:mma";
-                var time = match.valueForKey("alarmTime") as NSDate;
-                var timeString = format.stringFromDate(time);
-                
+                if(match.valueForKey("alarmTime") != nil){
+                    // prep date in specified format
+                    var format = NSDateFormatter();
+                    format.dateFormat = "hh:mma";
+                    var time = match.valueForKey("alarmTime") as NSDate;
+                    var timeString = format.stringFromDate(time);
+                    
+                    
+                    
+                    
+                    SetAlarm.text = timeString;
+                    if results.count > 1 {
+                        for(var i = 0; i < results.count - 1; i++){
+                            match = results[i] as NSManagedObject
+                            managedObjectContext?.deleteObject(match);
+                        }
+                        
+                    } else {
+                        println("break");
+                    }
+                }
                 // get stored military time and if it exists, set it to militaryTime variable
                 if(match.valueForKey("militaryTime") != nil){
-                    self.militaryTime = match.valueForKey("militaryTime") as Bool;                    
-                }
-                
-                
-                SetAlarm.text = timeString;
-                if results.count > 1 {
-                    for(var i = 0; i < results.count - 1; i++){
-                        match = results[i] as NSManagedObject
-                        managedObjectContext?.deleteObject(match);
-                    }
-                    
-                } else {
-                    println("break");
+                    self.militaryTime = match.valueForKey("militaryTime") as Bool;
                 }
             }
             
@@ -130,8 +136,10 @@ class FirstViewController: UIViewController {
         let alarm = Entity(entity: entityDescription!,
             insertIntoManagedObjectContext: managedObjectContext);
         
+        // save the alarm and military setting
         alarm.alarmTime = time;
         alarm.militaryTime = self.militaryTime;
+        
         managedObjectContext?.save(&error)
 
         if let err = error {
